training/src/data/requirements.txt
pandas~=2.2.3
lakefs~=0.7.1
Authlib~=1.3.2
confluent-kafka~=2.6.1
kafka-python-ng~=2.2.3
fsspec~=2024.10.0
lakefs-spec~=0.11.0
requests~=2.32.3